+++ docs/.vitepress/config.ts
@@ -113,6 +113,10 @@ let config = defineConfig({
                                 text: "@pitlane/data-table-d1",
                                 link: "/package/data-table-d1/index",
                             },
+                            {
+                                text: "@pitlane/data-table-d1/migrations",
+                                link: "/package/data-table-d1/migrations",
+                            },
                             { text: "@pitlane/dev", link: "/package/dev/index" },
                             {
                                 text: "@pitlane/dev/runtime",
@@ -136,6 +140,10 @@ let config = defineConfig({
         sidebar: {
             "/package/": [
                 { text: "@pitlane/data-table-d1", link: "/package/data-table-d1/" },
+                {
+                    text: "@pitlane/data-table-d1/migrations",
+                    link: "/package/data-table-d1/migrations",
+                },
                 { text: "@pitlane/dev", link: "/package/dev/" },
                 { text: "@pitlane/dev/runtime", link: "/package/dev/runtime" },
                 { text: "@pitlane/theme", link: "/package/theme/" },